Move argo tests
